fixed error due to circular import dependencies, had to move driver.py to pgtools utils
