+++ pgsmo/objects/database/database.py
@@ -5,7 +5,6 @@
 
 from typing import Optional               # noqa
 
-from pgsqltoolsservice.connection import ServerConnection    # noqa
 from pgsmo.objects.node_object import NodeCollection, NodeObject
 from pgsmo.objects.scripting_mixins import ScriptableCreate, ScriptableDelete
 from pgsmo.objects.server import server as s    # noqa
@@ -21,6 +20,7 @@ from pgsmo.objects.view.view import View
 from pgsmo.objects.view.materialized_view import MaterializedView
 from pgsmo.objects.extension.extension import Extension
 
+from pgsqltoolsservice.utils.driver import ServerConnection    # noqa
 
 class Database(NodeObject, ScriptableCreate, ScriptableDelete):
