+++ pgsmo/objects/server/server.py
@@ -6,14 +6,13 @@
 from typing import Dict, List, Mapping, Optional, Tuple, Callable      # noqa
 from urllib.parse import ParseResult, urlparse, quote_plus       # noqa
 
-
+from pgsqltoolsservice.utils.driver import ServerConnection
 from pgsmo.objects.node_object import NodeObject, NodeCollection, NodeLazyPropertyCollection
 from pgsmo.objects.database.database import Database
 from pgsmo.objects.role.role import Role
 from pgsmo.objects.tablespace.tablespace import Tablespace
 import pgsmo.utils as utils
 
-from pgsqltoolsservice.connection import ServerConnection
 
 class Server:
     TEMPLATE_ROOT = utils.templating.get_template_root(__file__, 'templates')
