+++ pgsqltoolsservice/connection/__init__.py
@@ -4,6 +4,5 @@
 # --------------------------------------------------------------------------------------------
 
 from pgsqltoolsservice.connection.connection_service import ConnectionInfo, ConnectionService
-from pgsqltoolsservice.connection.driver import DriverManager, ServerConnection
 
-__all__ = ['ConnectionInfo', 'ConnectionService', 'DriverManager', 'ServerConnection']
+__all__ = ['ConnectionInfo', 'ConnectionService']
